backend/requirements.txt
python-decouple==3.8
uvicorn[standard]
python-multipart==0.0.6
requests==2.28.2
fastapi==0.92.0
openai-whisper
